Refactored handler out of fragment
